+++ lib/src/yaml2podo.dart
@@ -117,8 +117,6 @@ T _toObject<T>(dynamic data, T Function(dynamic) fromJson) {
 
   Set<String> _usedMethods;
 
-  final String version = '0.1.11';
-
   bool _camelize;
 
   Map<String, _TypeInfo> _classes;